requirements.txt
boto3
fluent-logger